Se creo completamente los CRUDS con validaciones y se creo la compra y venta de productos

+++ src/minimercado/com/co/view/ProductView.java
@@ -11,7 +11,7 @@ import minimercado.com.co.utils.Misc;
 import java.util.Date;
 
 public class ProductView {
-  private static ProductController productController = new ProductController();
+  public ProductController productController = new ProductController();
 
   public void register() {
     char register;
@@ -20,12 +20,11 @@ public class ProductView {
       String name = Misc.getString("Ingrese el nombre: ");
       String description = Misc.getString("Ingrese la descripcion: ");
       Float price = Misc.getFloat("Ingrese el precio: ");
-      int quantity = Misc.getInteger("Ingrese la cantidad: ");
       String selectLineProduct= Misc.optionsSelector("Seleccione la linea de productos que vende: ", new String[]{"Linea de hogar", "Linea de aseo", "Linea de aseo personal", "Linea de fritas y verduras", "Linea de abarrotes", "Linea de confiteria"});
       LineProductController lineProductController = new LineProductController();
       LineProduct lineProduct = lineProductController.register(1, selectLineProduct);
 
-      boolean result = productController.register(code, name, description, price, quantity, lineProduct);
+      boolean result = productController.register(code, name, description, price, lineProduct);
       if (result) {
         Misc.showMessage("¡El producto se ha registrado exitosamente!");
       } else {
@@ -41,35 +40,36 @@ public class ProductView {
     String code = Misc.getString("Ingrese el numero de codigo del producto al cual desea consultar sus datos: ");
 
     String result = productController.search(code);
-    if (result != null && !result.isEmpty()) {
-      Misc.showMessage(result);
-    }
+    Misc.showMessage(result);
   }
 
   public void update() {
     String currentCode = Misc.getString("Ingrese el numero de codigo del producto al cual desea actualizar sus datos: ");
 
-    String search = productController.search(currentCode);
-    if (search.equals(currentCode)) {
-      Misc.showMessage("A continuacion ingrese los datos actualizados del producto");
+    if (!productController.list().startsWith("¡No hay")) {
+      if (!productController.search(currentCode).startsWith("¡Producto no ")) {
+        Misc.showMessage("A continuacion ingrese los datos actualizados del producto");
 
-      String code = Misc.getString("Ingrese el codigo: ");
-      String name = Misc.getString("Ingrese el nombre: ");
-      String description = Misc.getString("Ingrese la descripcion: ");
-      Float price = Misc.getFloat("Ingrese el precio: ");
-      int quantity = Misc.getInteger("Ingrese la cantidad: ");
-      String selectLineProduct= Misc.optionsSelector("Seleccione la linea de productos que vende: ", new String[]{"Linea de hogar", "Linea de aseo", "Linea de aseo personal", "Linea de fritas y verduras", "Linea de abarrotes", "Linea de confiteria"});
-      LineProductController lineProductController = new LineProductController();
-      LineProduct lineProduct = lineProductController.register(1, selectLineProduct);
+        String code = Misc.getString("Ingrese el codigo: ");
+        String name = Misc.getString("Ingrese el nombre: ");
+        String description = Misc.getString("Ingrese la descripcion: ");
+        Float price = Misc.getFloat("Ingrese el precio: ");
+        String selectLineProduct= Misc.optionsSelector("Seleccione la linea de productos a la que pertenece: ", new String[]{"Linea de aseo del hogar", "Linea de aseo personal", "Linea de fritas y verduras", "Linea de abarrotes", "Linea de confiteria"});
+        LineProductController lineProductController = new LineProductController();
+        LineProduct lineProduct = lineProductController.register(1, selectLineProduct);
 
-      boolean result = productController.register(code, name, description, price, quantity, lineProduct);
-      if (result) {
-        Misc.showMessage("¡El producto se ha registrado exitosamente!");
+        boolean result = productController.register(code, name, description, price, lineProduct);
+        if (result) {
+          Misc.showMessage("¡El producto se ha registrado exitosamente!");
+        } else {
+          Misc.showMessage("¡No se ha podido registrar el producto, intentelo nuevamente!");
+        }
       } else {
-        Misc.showMessage("¡No se ha podido registrar el producto, intentelo nuevamente!");
+        Misc.showMessage("¡Producto no encontrado, el numero de codigo ingresado no coincide con los datos guardados!");
       }
+
     } else {
-      Misc.showMessage("¡No hay registro de clientes!");
+      Misc.showMessage("¡No hay registro de productos!");
     }
   }
 
@@ -81,16 +81,18 @@ public class ProductView {
   public void delete() {
     String code = Misc.getString("Ingrese el numero de codigo del producto que desea eliminar: ");
 
-    String search = productController.search(code);
-    if (search == null && search.isEmpty()) {
-      Misc.showMessage("¡No hay productos para listar!");
-    }
-
-    if (productController.delete(code)) {
-      Misc.showMessage("¡Producto eliminado satisfactoriamente!");
+    if (!productController.list().startsWith("¡No hay")) {
+      if (!productController.search(code).startsWith("¡Producto no ")) {
+        if (productController.delete(code)) {
+          Misc.showMessage("¡Producto eliminado satisfactoriamente!");
+        } else {
+          Misc.showMessage("¡No se ha podido eliminar el producto, intentelo nuevamente!");
+        }
+      } else {
+        Misc.showMessage("¡Producto no encontrado, el numero de documento ingresado no coincide con los datos guardados!");
+      }
     } else {
-      Misc.showMessage("¡No se ha podido eliminar el producto!");
+      Misc.showMessage("¡No hay registro de productos!");
     }
-
   }
 }
